# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Stock Analytics Backend API
  version: 0.1.0
  description: API for PSX fundamental analytics platform
servers:
  - url: http://localhost:4000
    description: Local backend
paths:
...
  /api/companies:
    get:
      summary: Get companies list
      operationId: getCompanies
      parameters:
        - name: sector
          in: query
          description: Filter by sector
          schema:
            type: string
        - name: status
          in: query
          description: Listing status filter (e.g., NC)
          schema:
            type: string
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Limit number of results
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of companies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompaniesResponse'
...
components:
  schemas:
...
    CompaniesResponse:
      type: object
      properties:
        companies:
          type: array
          items:
            $ref: '#/components/schemas/CompanyBasic'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CompanyBasic:
      type: object
      properties:
        symbol:
          type: string
          example: HBL
        name:
          type: string
          example: Habib Bank Limited
        sector:
          type: string
          example: Banking
        marketCap:
          type: number
          description: Market cap in PKR millions
        price:
          type: number
          description: Current share price
        change:
          type: number
          description: Price change percentage
        pe:
          type: number
          description: P/E ratio
...